requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file> -c condo-forge
# platform: osx-64
backports=1.0=py_2
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
beautifulsoup4=4.11.1=pyha770c72_0
blas=1.0=mkl
bpemb=0.3.3=pyhd8ed1ab_0
brotli=1.0.9=h5eb16cf_7
brotli-bin=1.0.9=h5eb16cf_7
brotlipy=0.7.0=py39h63b48b0_1004
bzip2=1.0.8=h0d85af4_4
ca-certificates=2022.5.18.1=h033912b_0
certifi=2022.5.18.1=py39h6e9494a_0
cffi=1.15.0=py39he338e87_0
charset-normalizer=2.0.12=pyhd8ed1ab_0
click=8.1.3=py39h6e9494a_0
cloudpickle=2.1.0=pyhd8ed1ab_0
colorama=0.4.4=pyh9f0ad1d_0
conllu=4.4.2=pyhd8ed1ab_0
cryptography=37.0.2=py39h9c2a9ce_0
cycler=0.11.0=pyhd8ed1ab_0
dataclasses=0.8=pyhc8e2a94_3
deprecated=1.2.13=pyh6c4a22f_0
filelock=3.7.1=pyhd8ed1ab_0
fonttools=4.33.3=py39h701faf5_0
freetype=2.10.4=h4cff582_1
ftfy=6.1.1=pyhd8ed1ab_0
future=0.18.2=py39h6e9494a_5
gdown=4.4.0=pyhd8ed1ab_0
gensim=4.2.0=py39hd408605_0
giflib=5.2.1=hbcb3906_2
huggingface_hub=0.7.0=pyhd8ed1ab_0
hyperopt=0.2.7=pyhd8ed1ab_0
icu=70.1=h96cf925_0
idna=3.3=pyhd8ed1ab_0
importlib-metadata=4.11.4=py39h6e9494a_0
importlib_metadata=4.11.4=hd8ed1ab_0
intel-openmp=2022.0.0=hecd8cb5_3615
janome=0.4.2=pyhd8ed1ab_0
jinja2=3.1.2=pyhd8ed1ab_1
joblib=1.1.0=pyhd8ed1ab_0
jpeg=9e=h5eb16cf_1
kaleido-core=0.2.1=h0d85af4_0
kiwisolver=1.4.2=py39h7248d28_1
konoha=4.6.2=pyhd8ed1ab_1
langdetect=1.0.9=pyhd8ed1ab_0
lcms2=2.12=h577c468_0
lerc=3.0=he49afe7_0
libbrotlicommon=1.0.9=h5eb16cf_7
libbrotlidec=1.0.9=h5eb16cf_7
libbrotlienc=1.0.9=h5eb16cf_7
libcxx=14.0.4=hc203e6f_0
libdeflate=1.10=h0d85af4_0
libffi=3.4.2=h0d85af4_5
libgfortran=3.0.1=0
libiconv=1.16=haf1e3a3_0
libpng=1.6.37=h7cec526_2
libtiff=4.4.0=hfca7e8f_0
libwebp=1.2.2=h28dabe5_0
libwebp-base=1.2.2=h0d85af4_1
libxcb=1.13=h0d85af4_1004
libxml2=2.9.14=h08a9926_0
libxslt=1.1.35=heaa0ce8_0
libzlib=1.2.12=h6c3fc93_0
llvm-openmp=14.0.4=ha654fa7_0
lxml=4.9.0=py39h701faf5_0
lz4-c=1.9.3=he49afe7_1
markupsafe=2.1.1=py39h63b48b0_1
mathjax=2.7.7=0
matplotlib-base=3.5.2=py39h64a0072_0
mkl=2021.4.0=h89fa619_689
mkl-service=2.4.0=py39hadd8af7_0
mkl_fft=1.3.1=py39h7ae3660_1
mkl_random=1.2.2=py39h4d6be9b_0
more-itertools=8.13.0=pyhd8ed1ab_0
mpld3=0.5.8=pyhd8ed1ab_0
munkres=1.1.4=pyh9f0ad1d_0
ncurses=6.3=h96cf925_1
networkx=2.8.3=pyhd8ed1ab_0
ninja=1.11.0=h1b54a9f_0
nltk=3.7=pyhd3eb1b0_0
numpy=1.22.3=py39h2e5f0a9_0
numpy-base=1.22.3=py39h3b1a694_0
openjpeg=2.4.0=h6e7aa92_1
openssl=1.1.1o=hfe4f2af_0
overrides=3.0.0=pyh9f0ad1d_0
packaging=21.3=pyhd8ed1ab_0
pandas=1.4.2=py39hf72b562_2
pillow=9.1.1=py39h579eac4_1
pip=22.1.2=pyhd8ed1ab_0
plotly=5.8.1=pyhd8ed1ab_0
pptree=3.1=pyhd8ed1ab_0
pthread-stubs=0.4=hc929b4f_1001
py4j=0.10.9.5=pyhd8ed1ab_0
pycparser=2.21=pyhd8ed1ab_0
pyopenssl=22.0.0=pyhd8ed1ab_0
pyparsing=3.0.9=pyhd8ed1ab_0
pysocks=1.7.1=py39h6e9494a_5
python=3.9.13=h57e37ff_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python-flair=0.11.3=pyhd8ed1ab_0
python-kaleido=0.2.1=pyhd8ed1ab_0
python_abi=3.9=2_cp39
pytorch=1.10.2=cpu_py39h903acac_0
pytz=2022.1=pyhd8ed1ab_0
pyyaml=6.0=py39h63b48b0_4
readline=8.1=h05e3726_0
regex=2022.6.2=py39h701faf5_0
requests=2.27.1=pyhd8ed1ab_0
sacremoses=0.0.53=pyhd8ed1ab_0
scikit-learn=1.0.2=py39hae1ba45_1
scipy=1.7.3=py39h8c7af03_0
segtok=1.5.11=pyhd8ed1ab_0
sentencepiece=0.1.96=py39h7248d28_1
setuptools=62.3.3=py39h6e9494a_0
six=1.16.0=pyh6c4a22f_0
smart_open=6.0.0=pyhd8ed1ab_0
soupsieve=2.3.1=pyhd8ed1ab_0
sqlite=3.38.5=hd9f0692_0
sqlitedict=2.0.0=pyhd8ed1ab_0
tabulate=0.8.9=pyhd8ed1ab_0
tbb=2021.5.0=hbb4e6a2_1
tenacity=8.0.1=pyhd8ed1ab_0
threadpoolctl=3.1.0=pyh8a188c0_0
tk=8.6.12=h5dbffcc_0
tokenizers=0.12.1=py39h656764e_1
tqdm=4.64.0=pyhd8ed1ab_0
transformers=4.19.2=pyhd8ed1ab_0
typing-extensions=4.2.0=hd8ed1ab_1
typing_extensions=4.2.0=pyha770c72_1
tzdata=2022a=h191b570_0
unicodedata2=14.0.0=py39h63b48b0_1
urllib3=1.26.9=pyhd8ed1ab_0
wcwidth=0.2.5=pyh9f0ad1d_2
wheel=0.37.1=pyhd8ed1ab_0
wikipedia-api=0.5.4=pyhd8ed1ab_0
wrapt=1.14.1=py39h701faf5_0
xorg-libxau=1.0.9=h35c211d_0
xorg-libxdmcp=1.1.3=h35c211d_0
xz=5.2.5=haf1e3a3_1
yaml=0.2.5=h0d85af4_2
zipp=3.8.0=pyhd8ed1ab_0
zlib=1.2.12=h6c3fc93_0
zstd=1.5.2=ha9df2e0_1
